requirements.txt
django==3.2.4
gunicorn==20.1.0
django-heroku==0.3.1

djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.8.0

pre-commit==2.12.1
